fix #49 update chooser
